requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
pydantic==2.5.0
python-dotenv==1.0.0
numpy==1.26.2
streamlit==1.28.1
